Create prettybot-pull-request-automation.yml
PrettyBot is an automatic process for running a code formatting tool across entire codebases. It's kicked off on a schedule via a TeamCity build.

#project-miso-change-control has introduced a PRGB requirement for all changes to our software.

Although we are upping our change control policies, we still want PRs created by bots to auto-merge on a green build.

This PR uses the guidance here to introduce a GitHub Action workflow that auto approves PrettyBot PRs when they are created.

This will ensure PRs raised by the bot will auto-merge on a green build.